requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
backcall==0.2.0
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.8
comm==0.2.2
contourpy==1.3.0
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
datasets==3.2.0
dcase_util==0.2.20
debugpy==1.8.11
decorator==5.1.1
dill==0.3.8
distlib==0.3.9
dulwich==0.22.7
entrypoints==0.4
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
grpcio==1.68.0
huggingface-hub==0.26.3
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
installer==0.7.0
intervaltree==3.1.0
ipykernel==6.29.5
ipympl==0.9.6
ipython==8.12.0
ipywidgets==8.1.5
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jupyter-client==7.3.4
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keyring==25.6.0
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.9
llvmlite==0.43.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.2
pkginfo==1.12.0
platformdirs==4.3.6
poetry-core==2.0.0
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==5.29.0
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pybind11==2.13.6
pycparser==2.22
pydot-ng==2.0.0
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.0
pyproject_hooks==1.2.0
pyroomacoustics==0.7.7
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
SecretStorage==3.3.3
sed_eval==0.2.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
slugify==0.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
stack_data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.15
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchinfo==1.8.0
torchmetrics==1.6.0
torchvision==0.20.1
tornado==6.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
trove-classifiers==2025.1.6.15
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
virtualenv==20.28.1
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
